requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
requests
python-dotenv
celery
redis
django-celery-results